MOVE Render Related codes from startup_flow.py to render_control.py
