+++ release/scripts/startup/abler/export_tab/render_control.py
@@ -23,6 +23,7 @@ from ..lib import render, cameras
 from ..lib.materials import materials_handler
 from ..lib.tracker import tracker
 from time import time
+from ..warning_modal import BlockingModalOperator
 
 
 bl_info = {
@@ -85,6 +86,68 @@ class Acon3dCameraViewOperator(bpy.types.Operator):
         cameras.turn_on_camera_view()
 
         return {"FINISHED"}
+
+
+class Acon3dRenderWarningOperator(BlockingModalOperator):
+    bl_idname = "acon3d.render_warning"
+    bl_label = "Render Selected Scenes"
+    bl_description = "Render with high quality according to the set pixel"
+    bl_translation_context = "*"
+    scene_count: bpy.props.IntProperty(name="Scene count", default=0)
+
+    def draw_modal(self, layout):
+        padding_size = 0.01
+        content_size = 1.0 - 2 * padding_size
+        box = layout.box()
+        main = box.column()
+
+        main.label(text="")
+
+        row = main.split(factor=padding_size)
+        row.label(text="")
+        row = row.split(factor=content_size)
+        col = row.column()
+        col.label(text="Render selected scenes?")
+        col.label(text="High quality render may take a long time to be finished.")
+        col.label(text=f"Selected Scenes : {self.scene_count}")
+        if bpy.data.is_dirty:
+            col.operator("acon3d.render_save", text="Save and Render")
+        else:
+            col.operator("acon3d.render_high_quality", text="Render Selected Scenes")
+        col.operator("acon3d.close_blocking_modal", text="Cancel", text_ctxt="*")
+        row.label(text="")
+
+        main.label(text="")
+
+    @classmethod
+    def poll(self, context):
+        render_prop = context.window_manager.ACON_prop
+
+        is_method_selected = (
+            render_prop.hq_render_full
+            or render_prop.hq_render_line
+            or render_prop.hq_render_texture
+            or render_prop.hq_render_shadow
+        )
+        # 이 부분은 뺐다가 poll 실행 당시에 scene_count가 등록이 안되어있는 타이밍 문제때문에 에러메시지가 계속 떠서 그냥 다시 살림.
+        is_scene_selected = any(s.is_render_selected for s in render_prop.scene_col)
+        return is_method_selected and is_scene_selected
+
+
+def render_save_handler(dummy):
+    bpy.ops.acon3d.render_high_quality("INVOKE_DEFAULT")
+    bpy.app.handlers.save_post.remove(render_save_handler)
+
+
+class Acon3dRenderSaveOpertor(bpy.types.Operator):
+    bl_idname = "acon3d.render_save"
+    bl_label = "Save and Render"
+    bl_translation_context = "*"
+
+    def execute(self, context):
+        bpy.ops.acon3d.close_blocking_modal("INVOKE_DEFAULT")
+        bpy.app.handlers.save_post.append(render_save_handler)
+        return bpy.ops.acon3d.save("INVOKE_DEFAULT")
 
 
 class Acon3dRenderOperator(bpy.types.Operator):
@@ -626,6 +689,8 @@ class RenderProperty(bpy.types.PropertyGroup):
 classes = (
     RenderSceneInfoProperty,
     RenderProperty,
+    Acon3dRenderWarningOperator,
+    Acon3dRenderSaveOpertor,
     Acon3dCameraViewOperator,
     Acon3dRenderHighQualityOperator,
     Acon3dRenderSnipOperator,
